fix show picture on oled bug

+++ src/oled.h
@@ -40,13 +40,13 @@ void OLED_Clear(void);
 void OLED_DrawPoint(uint16_t x,uint16_t y,uint16_t t);
 void OLED_DrawLine(uint16_t x1,uint16_t y1,uint16_t x2,uint16_t y2,uint16_t mode);
 void OLED_DrawCircle(uint16_t x,uint16_t y,uint16_t r);
-void OLED_ShowChar(uint16_t x,uint16_t y,uint16_t chr,uint16_t size1,uint16_t mode);
+void OLED_ShowChar(uint16_t x,uint16_t y,uint8_t chr,uint16_t size1,uint16_t mode);
 void OLED_ShowChar6x8(uint16_t x,uint16_t y,uint16_t chr,uint16_t mode);
 void OLED_ShowString(uint16_t x,uint16_t y,uint8_t *chr,uint16_t size1,uint16_t mode);
 void OLED_ShowNum(uint16_t x,uint16_t y,uint32_t num,uint16_t len,uint16_t size1,uint16_t mode);
 void OLED_ShowChinese(uint16_t x,uint16_t y,uint16_t num,uint16_t size1,uint16_t mode);
 void OLED_ScrollDisplay(uint16_t num,uint16_t space,uint16_t mode);
-void OLED_ShowPicture(uint16_t x,uint16_t y,uint16_t sizex,uint16_t sizey,uint16_t BMP[],uint16_t mode);
+void OLED_ShowPicture(uint16_t x,uint16_t y,uint16_t sizex,uint16_t sizey,uint8_t BMP[],uint16_t mode);
 void OLED_Init(void);
 
 #endif
